+++ pydidas/gui/frames/data_browsing_frame.py
@@ -1,9 +1,11 @@
-# This file is part of pydidas.
+# This file is part of pydidas
+#
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -11,7 +13,7 @@
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
-# along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
+# along with pydidas If not, see <http://www.gnu.org/licenses/>.
 
 """
 Module with the DataBrowsingFrame which is used to browse through the filesystem in a
@@ -19,8 +21,8 @@ dedicated filesystem tree and show file data in a view window.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["DataBrowsingFrame"]
@@ -66,9 +68,6 @@ class DataBrowsingFrame(DataBrowsingFrameBuilder):
         self.sig_this_frame_activated.connect(self._widgets["viewer"].get_detector_size)
         self.sig_this_frame_activated.connect(
             partial(self._widgets["viewer"].cs_transform.check_detector_is_set, True)
-        )
-        self.sig_this_frame_activated.connect(
-            self._widgets["viewer"]._positionWidget.update_exp_setup_params
         )
 
     @QtCore.Slot(bool)