+++ pydidas/widgets/silx_plot/pydidas_position_info.py
@@ -1,9 +1,11 @@
-# This file is part of pydidas.
+# This file is part of pydidas
+#
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -11,7 +13,7 @@
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
-# along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
+# along with pydidas If not, see <http://www.gnu.org/licenses/>.
 
 """
 Module with the PydidasPositionInfo class used to display image data coordinates in
@@ -19,19 +21,18 @@ different coordinate systems.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["PydidasPositionInfo"]
 
 import numpy as np
 from qtpy import QtCore
-import pyFAI
 from silx.gui.plot.tools import PositionInfo
 
-from ...core.utils import pyfai_rot_matrix, get_chi_from_x_and_y
 from ...contexts import DiffractionExperimentContext
+from ...core.utils import get_chi_from_x_and_y
 
 
 AX_LABELS = {
@@ -55,6 +56,7 @@ class PydidasPositionInfo(PositionInfo):
             if diffraction_exp is None
             else diffraction_exp
         )
+        self._EXP.sig_params_changed.connect(self.update_exp_setup_params)
         self._x_widget = self.layout().itemAt(0).widget()
         self._y_widget = self.layout().itemAt(2).widget()
         self._cs_name = "cartesian"
@@ -115,32 +117,22 @@ class PydidasPositionInfo(PositionInfo):
     @QtCore.Slot()
     def update_exp_setup_params(self):
         """
-        Get the beam center in image coordinates.
-
-        Returns
-        -------
-        tuple
-            The beam center in a tuple (y, x)
+        Update beamcenter and detector distance from the DiffractionExperiment.
         """
-        poni = pyFAI.geometry.Geometry()
-
-        _theta = [self._EXP.get_param_value(f"detector_rot{dim}") for dim in [1, 2, 3]]
-        _y0 = self._EXP.get_param_value("detector_poni1")
-        _x0 = self._EXP.get_param_value("detector_poni2")
-        _z0 = self._EXP.get_param_value("detector_dist")
-        _rot = pyfai_rot_matrix(*_theta)
-        # the center is found by *subtracting* the rotation of the 0-position with
-        # respect to the poni (because pyFAI geometry moves the detector)
-        _beam_center = np.array([_y0, _x0, 0]) - np.dot(_rot, np.array([0, 0, _z0]))
-        _beam_center *= abs(_z0 / _beam_center[2])
-        self._beam_center = (_beam_center[0], _beam_center[1], _z0)
+        _f2dgeo = self._EXP.as_fit2d_geometry_values()
         self._pixelsize = (
             self._EXP.get_param_value("detector_pxsizex") * 1e-6,
             self._EXP.get_param_value("detector_pxsizey") * 1e-6,
         )
+        self._beam_center = (
+            _f2dgeo["center_y"],
+            _f2dgeo["center_x"],
+            _f2dgeo["det_dist"] * 1e-3,
+        )
 
     def _plotEvent(self, event):
-        """Handle events from the Plot.
+        """
+        Handle events from the Plot.
 
         :param dict event: Plot event
         """
@@ -185,8 +177,8 @@ class PydidasPositionInfo(PositionInfo):
         tuple
             The tuple with the polar r, chi coordinates.
         """
-        _x_rel = x_pix * self._pixelsize[0] - self._beam_center[1]
-        _y_rel = y_pix * self._pixelsize[1] - self._beam_center[0]
+        _x_rel = (x_pix - self._beam_center[1]) * self._pixelsize[0]
+        _y_rel = (y_pix - self._beam_center[0]) * self._pixelsize[1]
         _r = ((_x_rel) ** 2 + (_y_rel) ** 2) ** 0.5 * 1e3
         _chi = get_chi_from_x_and_y(_x_rel, _y_rel)
         return (_r, _chi)
@@ -207,8 +199,8 @@ class PydidasPositionInfo(PositionInfo):
         tuple
             The tuple with the polar 2-theta, chi coordinates.
         """
-        _x_rel = x_pix * self._pixelsize[0] - self._beam_center[1]
-        _y_rel = y_pix * self._pixelsize[1] - self._beam_center[0]
+        _x_rel = (x_pix - self._beam_center[1]) * self._pixelsize[0]
+        _y_rel = (y_pix - self._beam_center[0]) * self._pixelsize[1]
         _r = (_x_rel**2 + _y_rel**2) ** 0.5
         _2theta = np.arctan(_r / self._beam_center[2]) * 180 / np.pi
         _chi = get_chi_from_x_and_y(_x_rel, _y_rel)