+++ pydidas/widgets/silx_plot/silx_plot_stack_factory.py
@@ -70,10 +70,6 @@ def create_silx_plot_stack(
         frame.sig_this_frame_activated.connect(
             partial(frame._widgets["plot2d"].cs_transform.check_detector_is_set, True)
         )
-    if hasattr(frame, "sig_this_frame_activated"):
-        frame.sig_this_frame_activated.connect(
-            frame._widgets["plot2d"]._positionWidget.update_exp_setup_params
-        )
     frame.add_any_widget(
         "plot_stack",
         QtWidgets.QStackedWidget(),